Fix for Chrome 27+ being more strict about trying to access array elements that don't exist.
In the previous commit, I just checked to see if the user agent contained 'MSIE' with a version number of 9.0 or up.
It didn't cause any issues in Firefox or Chrome 25 on Linux, but Chrome 27 on windows was picky about accessing an
element that didn't exist (because there is no 'MSIE' string in the user agent). Now it checks Ext.isIE before trying
to get an MSIE version number out of the UA, so it should not cause any more problems.

Signed-off-by: Max Fierke <fierk007@umn.edu>